requirements.txt
numpy~=1.16.0
python-dateutil~=2.7.5
setuptools~=54.1.2
GitPython~=3.1.24